Allow exporting advanced search hits (#4988)
